+++ fns-hl7-pipeline/fn-hl7-json-lake/src/main/kotlin/gov/cdc/dex/hl7/Function.kt
@@ -9,7 +9,6 @@ import gov.cdc.dex.metadata.SummaryInfo
 import gov.cdc.dex.util.DateHelper.toIsoString
 import gov.cdc.dex.util.JsonHelper
 import gov.cdc.dex.util.JsonHelper.toJsonElement
-import gov.cdc.dex.util.ProcessingStatus.PSClientUtility
 import gov.cdc.hl7.bumblebee.HL7JsonTransformer
 import org.slf4j.LoggerFactory
 import java.util.*
@@ -50,10 +49,6 @@ class Function {
         }
         val outList = mutableListOf<String>()
         var returnValue = JsonObject()
-        val psClientUtility = PSClientUtility()
-        var traceId :String =""
-        var spanId :String =""
-        var childSpanId :String? =null
         try {
             messages.forEachIndexed { messageIndex: Int, singleMessage: String? ->
                 val startTime = Date().toIsoString()
@@ -61,21 +56,8 @@ class Function {
                     val inputEvent: JsonObject = JsonParser.parseString(singleMessage) as JsonObject
                     val filePath = JsonHelper.getValueFromJson("routing_metadata.ingested_file_path", inputEvent).asString
                     val messageUUID = JsonHelper.getValueFromJson("message_metadata.message_uuid", inputEvent).asString
-                    traceId = JsonHelper.getValueFromJson("routing_metadata.trace_id", inputEvent).asString
-                    spanId = JsonHelper.getValueFromJson("routing_metadata.span_id", inputEvent).asString
-
                     logger.info("DEX::Processing messageUUID:$messageUUID")
                     try {
-                        fnConfig.psURL?.let  {
-                            childSpanId =  psClientUtility.sendTraceToProcessingStatus(
-                                fnConfig.psURL,
-                                traceId,
-                                spanId,
-                                HL7JSONLakeStageMetadata.PROCESS_NAME
-                            )
-                            logger.info("Span ID of messageUUID: $messageUUID : ${childSpanId}")
-                        }
-
                         val hl7message = JsonHelper.getValueFromJsonAndBase64Decode("content", inputEvent)
                         val fullHL7WithNulls = buildJson(hl7message)
                         // remove nulls
@@ -101,20 +83,6 @@ class Function {
                         )
                         context.logger.info("Processed ERROR for HL7 JSON Lake messageUUID: $messageUUID, filePath: $filePath")
                     } // .catch
-                    finally { // closing the span for the msg
-                        fnConfig.psURL?.let {
-                            childSpanId?.let {
-                                psClientUtility.stopTrace(
-                                    fnConfig.psURL,
-                                    traceId,
-                                    it,
-                                    HL7JSONLakeStageMetadata.PROCESS_NAME
-                                )
-                            }
-                        }
-                    }
-
-
                 } catch (e: Exception) {
                     // message is bad, can't extract fields based on schema expected
                     context.logger.severe("Unable to process Message due to exception: ${e.message}")
@@ -141,6 +109,7 @@ class Function {
             logger.info("Send to event hub completed")
         } catch (e: Exception) {
             logger.error("Unable to send to event hub ${fnConfig.evHubSendName}: ${e.message}")
+            throw e
         }
         return returnValue
     }
@@ -166,18 +135,13 @@ class Function {
         logger.error(description)
         logger.info("Retrying send without output for message UUID $msgId")
         val msg = gsonWithNullsOn.toJson(message)
-        try {
-            val errors = fnConfig.evHubSender.send(msg)
-            if (errors.isEmpty()) {
-                logger.info("Second attempt successful for message UUID $msgId")
-            } else {
-                // not much else we can do
-                logger.error("SECOND ATTEMPT FAILED for message UUID $msgId. Message still too large.")
-            }
-        } catch (e: Exception) {
-            logger.error("Unable to send to event hub ${fnConfig.evHubSendName}: ${e.message}")
+        val errors = fnConfig.evHubSender.send(msg)
+        if (errors.isEmpty()) {
+            logger.info("Second attempt successful for message UUID $msgId")
+        } else {
+            // not much else we can do
+            logger.error("SECOND ATTEMPT FAILED for message UUID $msgId. Message still too large.")
         }
-
     }
 
     private fun updateMetadataAndDeliver(