+++ fns-hl7-pipeline/fn-redactor/src/main/kotlin/gov/cdc/dex/hl7/FunctionConfig.kt
@@ -9,8 +9,7 @@ class FunctionConfig {
     val profileConfig : ProfileConfiguration
     val evHubSender: DedicatedEventHubSender
     val evHubSendName: String = System.getenv("EventHubSendName")
-    val psURL = System.getenv("ProcessingStatusBaseURL")
-
+    
     init {
          //Init Event Hub connections
         val evHubConnStr = System.getenv("EventHubConnectionString")